Lab3/rest/requirements/base.txt
django==1.9.*
djangorestframework==3.4.*
django-filter==0.13.*
django-cors-headers==1.1.*
bcrypt==3.1.*
psycopg2==2.6.*
pytz==2016.6.*
django-import-export==0.4.*
django-dirtyfields==1.1.*
requests==2.10.*
celery[redis]==3.1.*
django-embed-video==1.1.*
goose-extractor==1.0.*
django_countries==3.4.*
django-rest-auth[with_social]==0.8.*
xlsxwriter==0.9.3
xlrd
schedule
beautifulsoup4
urlparse2
